pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openapi2mcp"
version = "0.1.0"
description = "A tool to convert OpenAPI specs to MCP server with tools"
readme = "README.md"
authors = [
    {name = "OpenAPI2MCP Contributors"}
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.2",
    "pyyaml>=6.0",
    "pydantic>=2.3.0",
    "openapi3-parser>=1.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=1.6.5"
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "pytest>=7.4.0",
]

[project.scripts]
openapi2mcp = "openapi2mcp.cli:main"
